C++AL: - Reorganized source code for Eigen integration - Split up source code into implementation and prototype header - Doxygen improvements. Documentation now works again, also with Doxygen 1.7.6.1. Worked around deficiencies in Doxygen's preprocessor
