+++ src/ports/greenplum/dbconnector/Compatibility.cpp
@@ -1,6 +1,6 @@
 /* ----------------------------------------------------------------------- *//**
  *
- * @file GPCompatibility.cpp
+ * @file greenplum/dbconnector/Compatibility.cpp
  *
  *//* ----------------------------------------------------------------------- */
