+++ src/ports/postgres/dbconnector/Allocator_proto.hpp
@@ -4,23 +4,25 @@
  *
  *//* ----------------------------------------------------------------------- */
 
+// Workaround for Doxygen: Ignore if not included by dbconnector.hpp
+#ifdef MADLIB_DBCONNECTOR_HPP
+
 /**
  * @brief PostgreSQL memory allocator
  *
- * The main issue consideration for memory allocation in PostgreSQL is:
- * -# Memory leaks are only guaranteed not to occur if PostgreSQL memory
- *    allocation functions are used\n
- *    .
- *    PostgreSQL knows the concept of "memory contexts" such as current function
- *    call, current aggregate function, or current transaction. Memory
- *    allocation using \c palloc() always occurs within a specific memory
- *    context -- and once a memory context goes out of scope all memory
- *    associated with it will be deallocated (garbage collected).\n
- *    .
- *    As a second safety measure we therefore redefine <tt>operator new</tt> and
- *    <tt>operator delete</tt> to call \c palloc() and \c pfree(). (This is
- *    essentially an \b additional protection against leaking C++ code. Given
- *    1., no C++ destructor call will ever be missed.)
+ * PostgreSQL knows the concept of "memory contexts" such as current function
+ * call, current aggregate function, or current transaction. Memory
+ * allocation using \c palloc() always occurs within a specific memory
+ * context -- and once a memory context goes out of scope all memory
+ * associated with it will be deallocated (garbage collected).
+ * 
+ * In C++ code one should in general not rely on this form of garbage
+ * collection, as destructors could be used for releasing other resources
+ * than memory. Nonetheless, protection at least against memory leaks is better
+ * than no protection.
+ * 
+ * In NewDelete.cpp, we therefore redefine <tt>operator new()</tt> and
+ * <tt>operator delete()</tt> to call \c palloc() and \c pfree().
  *
  * @see AbstractionLayer::Allocator::internalAllocate, NewDelete.cpp
  *
@@ -29,20 +31,15 @@
  */
 class AbstractionLayer::Allocator {
 public:
-    enum ReallocateMemory {
-        NewAllocation,
-        Reallocation
-    };    
-
     Allocator(FunctionCallInfo inFCInfo) : fcinfo(inFCInfo) { }
-
-    template <typename T>
-    MutableArrayHandle<T> allocateArray(size_t inNumElements) const;
 
     template <typename T, dbal::MemoryContext MC,
         dbal::ZeroMemory ZM, dbal::OnMemoryAllocationFailure F>
     MutableArrayHandle<T> allocateArray(size_t inNumElements) const;
     
+    template <typename T>
+    MutableArrayHandle<T> allocateArray(size_t inNumElements) const;
+
     template <dbal::MemoryContext MC, dbal::ZeroMemory ZM,
         dbal::OnMemoryAllocationFailure F>
     void *allocate(const size_t inSize) const;
@@ -55,6 +52,14 @@ public:
     void free(void *inPtr) const;
 
 protected:
+    /**
+     * @brief Template argument type for internalAllocate()
+     */
+    enum ReallocateMemory {
+        NewAllocation,
+        Reallocation
+    };
+
     template <dbal::ZeroMemory ZM>
     void *internalPalloc(size_t inSize) const;
     
@@ -68,5 +73,14 @@ protected:
         dbal::OnMemoryAllocationFailure F, Allocator::ReallocateMemory R>
     void *internalAllocate(void *inPtr, const size_t inSize) const;
 
+    /**
+     * @brief The PostgreSQL FunctionCallInfo passed to the UDF
+     *
+     * @internal
+     *     The name \c fcinfo is choosen on purpose because several PostgreSQL
+     *     macros rely on it.
+     */
     FunctionCallInfo fcinfo;
 };
+
+#endif // MADLIB_DBCONNECTOR_HPP (workaround for Doxygen)
