+++ src/ports/postgres/dbconnector/OutputStreamBuffer_proto.hpp
@@ -4,6 +4,9 @@
  *
  *//* ----------------------------------------------------------------------- */
 
+// Workaround for Doxygen: Ignore if not included by dbconnector.hpp
+#ifdef MADLIB_DBCONNECTOR_HPP
+
 /**
  * @brief Stream buffer that dispatches all output to PostgreSQL's ereport
  *        function
@@ -15,3 +18,5 @@ class AbstractionLayer::OutputStreamBuffer
 public:
     void output(char *inMsg, uint32_t inLength) const;
 };
+
+#endif // MADLIB_DBCONNECTOR_HPP (workaround for Doxygen)
