+++ src/ports/postgres/dbconnector/PGException_proto.hpp
@@ -4,6 +4,9 @@
  *
  *//* ----------------------------------------------------------------------- */
 
+// Workaround for Doxygen: Ignore if not included by dbconnector.hpp
+#ifdef MADLIB_DBCONNECTOR_HPP
+
 /**
  * @brief Unspecified PostgreSQL backend expcetion
  */
@@ -15,3 +18,5 @@ public:
     PGException()
       : std::runtime_error("The backend raised an exception.") { }
 };
+
+#endif // MADLIB_DBCONNECTOR_HPP (workaround for Doxygen)
